requirements.txt
beautifulsoup4==4.12.3
docx==0.2.4
EbookLib==0.18
fitz==0.0.1.dev2
numpy==1.24.4
pandas==1.5.3
psutil==5.9.5
python_docx==1.1.2
python_pptx==0.6.23
PyYAML==6.0.1
scikit_learn==1.5.0
sentence_transformers==2.2.2
setuptools==68.0.0
torch==2.2.2
tqdm==4.66.1
transformers==4.41.2
